Feedback: Sunrise
This is a very good effort. I really like the approach taken to sun motion. This is much more efficent than the trig maths. Unfortunately at the pixel level the "world" is too small for acceleration to take effect. You need to create a larger virtual screen and the shrink everything down. Not as hard as it sounds.